[refactor] 패키징화 + 리스트팝업 프리뷰 추가
